Creation of App\Mailer in order to keep Core\Mailer generic
