+++ core/Service/Mailer.php
@@ -9,7 +9,7 @@
 namespace Core\Service;
 
 
-class Mailer
+abstract class Mailer
 {
     private $mailer;
 
@@ -17,24 +17,7 @@ class Mailer
         $this->setMailer($mailer);
     }
 
-    public function send($sender_name, $sender_mail, $mail_content){
-        $message = (new \Swift_Message($sender_name . ' cherche à te joindre'))
-            ->setFrom([$sender_mail => $sender_name])
-            ->setTo(['thiblaf10@gmail.com' => 'Thibaud Lafont'])
-            ->setBody("
-                / Infos sur l'expéditeur /
-                
-                Nom  : {$sender_name} 
-                Mail : {$sender_mail}
-                
-                
-                / Contenu /
-                
-                $mail_content
-            ");
-
-        $this->getMailer()->send($message);
-    }
+    public abstract function send($sender_name, $sender_mail, $mail_content);
 
     public function getMailer(){
         return $this->mailer;
